+++ CHANGELOG.md
@@ -6,6 +6,7 @@ This file contains a list of changes for each released version.
 - Core engine rewritten in GO for portability and maintainability.
 - Repository history cleaned to remove legacy implementation artifacts.
 - Port List and List Builder to GO.
+- Port Docker function to GO.
 
 ## v0.10.0
 - Bring in the wrapper